refactor(website): Rename files
