+++ website/index.html
@@ -171,7 +171,7 @@
       <p><a href="#">Back to top</a></p>
     </footer>
 
+    <script src="src/placeholders.js" defer></script>
     <script src="src/index.js" defer></script>
-    <script src="src/app.js" defer></script>
   </body>
 </html>
