+++ website/src/index.js
@@ -1,11 +1,47 @@
-// Show placeholders for paragraphs
-const paragraphs = [].slice.call(document.querySelectorAll('p.placeholder'))
+import mediumZoom from 'medium-zoom'
 
-paragraphs.forEach(paragraph => {
-  // eslint-disable-next-line no-param-reassign
-  paragraph.innerHTML = paragraph.textContent
-    .split(' ')
-    .filter(text => text.length > 4)
-    .map(text => `<span class="placeholder__word">${text}</span>`)
-    .join(' ')
+const zoomDefault = mediumZoom('#zoom-default')
+const zoomMargin = mediumZoom('#zoom-margin', { margin: 48 })
+const zoomBackground = mediumZoom('#zoom-background', { background: '#212530' })
+const zoomScrollOffset = mediumZoom('#zoom-scrollOffset', {
+  scrollOffset: 0,
+  background: 'rgba(25, 18, 25, .9)',
+})
+
+// Trigger the zoom when the button is clicked
+const zoomToTrigger = mediumZoom('#zoom-trigger')
+const button = document.querySelector('#button-trigger')
+button.addEventListener('click', () => zoomToTrigger.show())
+
+// Detach the zoom after having been zoomed once
+const zoomToDetach = mediumZoom('#zoom-detach')
+zoomToDetach.addEventListeners('hidden', () => zoomToDetach.detach())
+
+// Observe zooms to write the history
+const observedZooms = [
+  zoomDefault,
+  zoomMargin,
+  zoomBackground,
+  zoomScrollOffset,
+  zoomToTrigger,
+  zoomToDetach,
+]
+
+// Log all interactions in the history
+const history = document.querySelector('#history')
+
+observedZooms.forEach(zoom => {
+  zoom.addEventListeners('show', event => {
+    const time = new Date().toLocaleTimeString()
+    history.innerHTML += `<li>Image "<em>${
+      event.target.alt
+    }</em>" was zoomed at ${time}</li>`
+  })
+
+  zoom.addEventListeners('detach', event => {
+    const time = new Date().toLocaleTimeString()
+    history.innerHTML += `<li>Image <em>"${
+      event.target.alt
+    }"</em> was detached at ${time}</li>`
+  })
 })
